backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.3
celery==5.3.6
psycopg2-binary==2.9.9
pytest-django==4.8.0
locust==2.24.0
daphne==4.1.0
websockets==12.0
locust==2.24.0
websocket-client==1.7.0
dj-database-url>=2.0
django-cors-headers==4.3.1
gevent==23.9.1
